Filter for number of clusters added
I added a function which checks if the number of clusters exists and filters out all clusterings with that number of clusters. If the number doesn't exist it takes the two numbers of clusters closest to the given number. If the number of clusters exists only once it takes the two closest again. 
I have now tried to add a method which merges vertices with edge weight 1 between them. But I don't understand the mapping which is required by the igraph method contract_vertices. Is there some documentation or code example for python, I only found a code example for R.
Also some changes in display, added a method for analyzing but still in progress and added variabel for showing the labels in the plot.