merge_weight_1 function added
Added the merge_weight_1 function in the graph analysis. I think it works but we should add more testing. As the mapping in the iGraph contract_vertices function is a bit tricky and I generate nodes without values while merging as all possible nodes, which have to be deleted later on. 
Added the docstrings and also generalized the plotting so there is a function, which can be called to plot the graphs at any point.