fix the bug while merging edges
Replaced the merge_edges_weight_above_threshold function with a new one using the iGraph vertex clustering to merge the edges. Also added two functions providing calculations for the function, one from stackoverflow calculating set intersections as this was buggy in my version.

Wrote two more tests to check if everything works.

Changed the merge_by_list to be used in combination with the merge edges function.

Changed the save_and_load_graphs to file names as this is more practical.

And recalculate the edges on final time in the merge_resolution_graphs as the probability_df is not updated after the last merger.